Make first test of location model work.

+++ db/migrate/20150514230218_create_locations.rb
@@ -2,8 +2,8 @@ class CreateLocations < ActiveRecord::Migration
 	def change
 		create_table :locations do |t|
 
-			t.decimal :lat
-			t.decimal :lng
+			t.integer :lat
+			t.integer :lng
 
 			t.timestamps null: false
 		end
